Deleted user and moved functions to account
