+++ app/src/main/java/ch/epfl/sweng/SDP/home/HomeActivity.java
@@ -26,8 +26,6 @@ import ch.epfl.sweng.SDP.firebase.Database;
 import ch.epfl.sweng.SDP.game.VotingPageActivity;
 import ch.epfl.sweng.SDP.game.WaitingPageActivity;
 
-import ch.epfl.sweng.SDP.matchmaking.User;
-
 
 import com.bumptech.glide.Glide;
